refactor: перемещены файлы js в папки в scripts
