+++ index.html
@@ -124,6 +124,6 @@
     </li>
   </template>
 
-  <script type="module" src="scripts/index.js"></script>
+  <script type="module" src="scripts/pages/index.js"></script>
 </body>
 </html>
