+++ scripts/utils/constants.js
@@ -1,7 +1,7 @@
 'use strict';
 
 
-import FormValidator from './FormValidator.js';
+import FormValidator from '../components/FormValidator.js';
 
 export
 const formSetting = {
